
\name{Judge3}
\alias{Judge3}
\docType{data}
\title{
GPCSIV, dataset Judge 3
}
\description{
This  is a simulated dataset which contains informations about   assessed by 3 experts to evaluate various product from six regions. 
This dataset (Judge1) contains 4 symbolic interval variables. 
In these data, the five successive pairs of columns represented the five symbolic variables of type interval.
Each pair consists of a min and a max.
}
\usage{data(Judge1)}
\format{
  A data frame with 6 observations on the following 8 variables.
  \describe{
    \item{\code{Banana.min}}{a numeric vector, minimum grade banana proposed by the expert 3}
    \item{\code{Banana.max}}{a numeric vector, maximum grade banana proposed by the expert 3}
    \item{\code{Coffee.min}}{a numeric vector, minimum grade coffee proposed by the expert 3}
    \item{\code{Coffee.max}}{a numeric vector, maximum grade coffee proposed by the expert 3}
    \item{\code{Thea.min}}{a numeric vector, minimum grade thea proposed by the expert 3}
    \item{\code{Thea.max}}{a numeric vector, maximum grade thea proposed by the expert 3}
    \item{\code{Cocoa.min}}{a numeric vector, minimum grade cocoa proposed by the expert 3}
    \item{\code{Cocoa.max}}{a numeric vector, maximum grade banana proposed by the expert 3}
  }
}

%%\references{
%%  ~~ possibly secondary sources and usages ~~
%%}

\examples{
data(Judge3)
}

